Move search into resources
